Making redirect tests check for signature validity (that fails)
